Extract serverless agent initialization
